fix: Backend incorrect column type and changing JWT as env variable and enforcing architecture pattern

+++ backend/src/main/java/com/g9latam/team62/fintech_api/dto/UserResponseDTO.java
@@ -13,7 +13,7 @@ public record UserResponseDTO(
         BigDecimal monthlyIncome,
         SavingFrequency savingFrequency,
         FinancialProfile financialProfile,
-        BigDecimal profileAccuracy
+        Double profileAccuracy
 ) {
     public static UserResponseDTO fromEntity(User user) {
         if (user == null) {
